bar chart simplification

+++ src/main/java/edu/wright/hendrix11/svg/jsf/SvgJsfComponent.java
@@ -12,8 +12,6 @@
 
 package edu.wright.hendrix11.svg.jsf;
 
-import edu.wright.hendrix11.svg.jsf.barChart.BarChartModel;
-
 import javax.faces.component.UIComponentBase;
 
 /**
@@ -60,7 +58,7 @@ public abstract class SvgJsfComponent extends UIComponentBase
     /**
      * @param height
      */
-    public void setHeight(Integer height)
+    public void setHeight(Double height)
     {
         getStateHelper().put(PropertyKeys.height, height);
     }
@@ -68,7 +66,7 @@ public abstract class SvgJsfComponent extends UIComponentBase
     /**
      * @param height
      */
-    public void setHeight(Double height)
+    public void setHeight(Integer height)
     {
         getStateHelper().put(PropertyKeys.height, height);
     }
@@ -84,7 +82,7 @@ public abstract class SvgJsfComponent extends UIComponentBase
     /**
      * @param width
      */
-    public void setWidth(Integer width)
+    public void setWidth(Double width)
     {
         getStateHelper().put(PropertyKeys.width, width);
     }
@@ -92,7 +90,7 @@ public abstract class SvgJsfComponent extends UIComponentBase
     /**
      * @param width
      */
-    public void setWidth(Double width)
+    public void setWidth(Integer width)
     {
         getStateHelper().put(PropertyKeys.width, width);
     }
